+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- Removed *quote* prefix from QuoteHandler method names, for better chaining
+  from Adapter, eg. `$adapter->quote()->into()`
 ### Removed
 - Drop support for PHP 5.5
 - Helpers have been migrated to the `phlib/db-helper` package: `BulkInsert`,
@@ -12,7 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Remove `QuoteableInterface` and `QuoteableAdapterInterface`
 - Remove QuoteHandler pass-through and setter methods from Adapter. Instead use
   QuoteHandler directly by chaining, eg. `$adapter->quoteInto()` is replaced
-  with `$adapter->quote()->quoteInto()`
+  with `$adapter->quote()->into()`
 
 ## [0.0.5] - 2017-01-03
 ### Fixed